requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pandas>=2.2.0
plotly>=5.18.0
pytest>=8.0.0


